Feat : 파일 사이즈 제한 없애기

+++ src/main/java/project/backend/domain/member/dto/MemberStatisticsResponseDto.java
@@ -9,6 +9,6 @@ import lombok.*;
 @AllArgsConstructor
 public class MemberStatisticsResponseDto {
     private String category;
-    private Integer categoryCnt;
+    private Long categoryCnt;
     private Double categoryPercent;
 }
